requirements.txt
streamlit==1.36.0
python-dotenv==1.0.1
gtts==2.5.3
pysqlite3-binary==0.5.3

# LangChain 0.2.x (Stable + LCEL)
langchain-core==0.2.12
langchain==0.2.12
langchain-community==0.2.12
langchain-groq==0.1.3
langchain-huggingface==0.0.3

# Vector Store
chromadb==0.5.3

# Embeddings
sentence-transformers==2.7.0
torch==2.1.2
transformers==4.40.0
accelerate==0.28.0
huggingface-hub==0.23.0
numpy==1.26.4